backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
opencv-python-headless
numpy
Pillow
ImageHash
torch
torchvision
supabase
python-dotenv
